- added class AccountDefinitionGroup and Definitions

+++ src/main/java/com/soramusoka/destinyApiClient/dto_layer/account_summary/CharacterBase.java
@@ -13,8 +13,8 @@ public class CharacterBase {
     public double raceHash;
     public double genderHash;
     public double classHash;
-    public int currentActivityHash;
-    public int lastCompletedStoryHash;
+    public double currentActivityHash;
+    public double lastCompletedStoryHash;
     public int grimoireScore;
     public int genderType;
     public int classType;
